requirements.txt
fastapi==0.108.0
langchain==0.1.0
langserve==0.0.39
python-dotenv==1.0.0
uvicorn==0.25.0
